rename icons with Icon prefix
