+++ src/lib/components/icons/IconCopy.svelte
